rm unneeded files
