--- src/components/Home.jsx
@@ -1,15 +1,0 @@
-import Hero from "./Landing_Page/hero";
-import Features from "./Landing_Page/features";
-import Footer from "./Landing_Page/footer";
-
-function Home() {
-  return (
-    <>
-      <Hero />
-      <Features />
-      <Footer />
-    </>
-  );
-}
-
-export default Home;
