+++ src/root.jsx
@@ -31,10 +31,10 @@ export default function Root() {
           href="https://fonts.gstatic.com"
           crossorigin=""
         />
-        {/* <link
+        <link
           href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700;800&display=swap"
           rel="stylesheet"
-        /> */}
+        />
         {/* meta description */}
         <Title>NvChad</Title>
         <Meta charset="utf-8" />
