+++ src/routes/(index)/index.jsx
@@ -1,3 +1,14 @@
-import Home from "~/components/Home";
+import Hero from "~/components/Landing_Page/hero";
+import Features from "~/components/Landing_Page/features";
+import Footer from "~/components/Landing_Page/footer";
 
+function Home() {
+  return (
+    <>
+      <Hero />
+      <Features />
+      <Footer />
+    </>
+  );
+}
 export default Home;
